+++ app/Policies/MachinePolicy.php
@@ -5,6 +5,7 @@ namespace App\Policies;
 use App\User;
 use App\Machine;
 use Illuminate\Auth\Access\HandlesAuthorization;
+use Illuminate\Auth\Access\Response as AccessResponse;
 use Illuminate\Support\Facades\DB;
 
 /**
@@ -31,6 +32,19 @@ class MachinePolicy
     }
 
     /**
+     * Pre-authorization checks: Always allow roles which hold the `global` action (usually the admin role).
+     * 
+     * @param User $user
+     * @param string $ability
+     * @return bool|null
+     */
+    public function before(User $user, string $ability): bool|null
+    {
+        if ($user->isAdmin()) return true;
+        return null;
+    }
+
+    /**
      * Determine whether the user can view any Machines.
      *
      * With business units disabled:
@@ -39,17 +53,17 @@ class MachinePolicy
      * With business units enabled:
      * - Admins can always view.
      * - Manager or User can only view their own Business Unit.
-     * - Nobody role cannot view anything.
+     * - Nobody role cannot view anything. (Nobody is a user who is part of munkireport but belongs to no business units)
      *
      *
      * @param User $user
      * @return mixed
      */
-    public function viewAny(User $user)
+    public function viewAny(User $user): AccessResponse
     {
         // user must be admin or manager
         if (!config('_munkireport.enable_business_units', false)) return true;
-        if ($user->isAdmin()) return true;
+
 
         // TODO: scan business unit membership
 
@@ -68,13 +82,16 @@ class MachinePolicy
      * @param Machine $machine
      * @return mixed
      */
-    public function view(User $user, Machine $machine)
+    public function view(User $user, Machine $machine): AccessResponse
     {
-        // user must be admin or manager
+        // there are no view restrictions when BU is turned off
         if (!config('_munkireport.enable_business_units', false)) return true;
-        if ($user->isAdmin()) return true;
 
-        // TODO: scan business unit membership
+        // Determine whether the user can view based on their machine groups which was decided by the MachineGroupMembership Auth Listener
+        $machineGroups = session()->get('machine_groups', []);
+        $matchMachineGroup = $machine->reportData()->firstOrFail()->machine_group;
+        if (in_array((string)$matchMachineGroup, $machineGroups)) return true;
+
 
         return false;
     }
@@ -82,35 +99,35 @@ class MachinePolicy
     /**
      * Determine whether the user can delete the model.
      *
+     * With business units enabled:
+     * - You need to be a manager in the same business unit *OR* a global admin.
+     * With business units disabled:
+     * - Be a global manager or admin
+     *
      * @param User $user
      * @param Machine $machine
      * @return mixed
      */
-    public function delete(User $user, Machine $machine)
+    public function delete(User $user, Machine $machine): AccessResponse
     {
-        if ($user->isAdmin()) return true;
+        if (!config('_munkireport.enable_business_units', false) && $user->isManager()) return true;
 
-        $managerOfBusinessUnits = $user->managerOfBusinessUnits()->get('business_units.id');
-
-        $managesMachineBusinessUnit = DB::table('machine')
+        $machineGroupBusinessUnits = DB::table('machine')
             ->join(
                 'reportdata',
                 'machine.serial_number', '=', 'reportdata.serial_number'
             )
             ->join(
-                'machine_groups',
-                'reportdata.machine_group', '=', 'machine_groups.id'
+                'machine_group',
+                'reportdata.machine_group', '=', 'machine_group.groupid'
             )
             ->join(
-                'business_units',
-                'machine_groups.business_unit_id', '=', 'business_units.id'
-            )
-            ->where('machine.id', '=', $machine->id)
-            ->whereIn('business_units.id', $managerOfBusinessUnits);
+                'business_unit', function (JoinClause $join) {
+                    $join->on('reportdata.machine_group', '=', 'business_unit.value')
+                        ->where('business_unit.property', '=', 'machine_group');
+                })
+            ->get();
 
-        if ($managesMachineBusinessUnit->count() > 0) {
-            return true;
-        }
 
         return false;
     }