+++ database/factories/UserFactory.php
@@ -3,7 +3,7 @@
 namespace Database\Factories;
 
 use App\Models\Team;
-use App\Models\User;
+use App\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 use Laravel\Jetstream\Features;
@@ -25,17 +25,46 @@ class UserFactory extends Factory
     public function definition(): array
     {
         return [
-            'name' => $this->faker->name(),
-            'email' => $this->faker->unique()->safeEmail(),
+            'name' => fake()->name(),
+            'email' => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
-            'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+            'password' => bcrypt('password'),
             'two_factor_secret' => null,
             'two_factor_recovery_codes' => null,
+            'role' => 'user',
             'remember_token' => Str::random(10),
+            'source' => null,
+            'display_name' => fake()->name(),
             'profile_photo_path' => null,
             'current_team_id' => null,
+            'locale' => 'en_US',
         ];
     }
+
+    /**
+     * Indicate that the user should have the global admin role
+     */
+    public function admin(): static
+    {
+        return $this->state(function (array $attributes) {
+           return [
+               'role' => 'admin',
+           ];
+        });
+    }
+
+    /**
+     * Indicate that the user should have the global manager role
+     */
+    public function manager(): static
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'role' => 'manager',
+            ];
+        });
+    }
+
 
     /**
      * Indicate that the model's email address should be unverified.