+++ resources/js/Pages/Dashboard.vue
@@ -4,5 +4,6 @@ import Welcome from '@/Components/Welcome.vue';
 </script>
 
 <template>
+  <div />
   <!-- This is the default landing space of Jetstream which we do not use -->
 </template>