+++ resources/views/layouts/unauthenticated.blade.php
@@ -64,7 +64,7 @@ It may be deleted when all the login/logout/password views are migrated.
     @endif
 
 </head>
-<body class="mr-hybrid-layout" style="padding-top: 56px;">
+<body class="mr-unauthenticated-layout" style="padding-top: 56px;">
 <nav class="navbar navbar-expand-lg navbar-light bg-light fixed-top">
     <a class="navbar-brand" href="{{ url('/') }}">{{ config('app.name', 'MunkiReport') }}</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">