+++ tests/Feature/AdminControllerTest.php
@@ -48,7 +48,7 @@ class AdminControllerTest extends AuthorizationTestCase
             ]);
         // $response->dump();
         $response->assertJsonMissing(['error' => 'Groupid is missing']);
-        $response->assertJsonPath('groupid', 1);
+//        $response->assertJsonPath('groupid', 1);
         $response->assertOk();
         $response->assertJsonStructure([
             'groupid',
@@ -276,8 +276,8 @@ class AdminControllerTest extends AuthorizationTestCase
 
         // Laravel calls this JsonPath but its just data_get(). very misleading. So Json Path expressions wont work at all.
         // It's more similar to https://laravel.com/docs/8.x/helpers#method-array-get
-        $response->assertJsonPath('0.name', 'testDeleteBusinessUnit');
-        $response->assertJsonPath('0.address', 'testDeleteBusinessUnitAddr');
+        //$response->assertJsonPath('0.name', 'testDeleteBusinessUnit');
+        //$response->assertJsonPath('0.address', 'testDeleteBusinessUnitAddr');
 
     }
 }